+++ tests/test_helpers.py
@@ -120,11 +120,32 @@ def test_vcard_title_escape():
     assert 'BEGIN:VCARD\r\nVERSION:3.0\r\nN:Doe;John\r\nFN:John Doe\r\nTITLE:Director\, Research and Development\r\nEND:VCARD\r\n' == vcard
 
 
+def test_vcard_data_valid_bday():
+    expected_vcard_data = 'BEGIN:VCARD\r\nVERSION:3.0\r\nN:Mustermann;Max\r\nFN:Max Mustermann\r\nBDAY:1976-09-19\r\nEND:VCARD\r\n'
+    vcard = helpers.make_vcard_data('Mustermann;Max', 'Max Mustermann', birthday='1976-09-19')
+    assert expected_vcard_data == vcard
+    vcard = helpers.make_vcard_data('Mustermann;Max', 'Max Mustermann', birthday=date(year=1976, month=9, day=19))
+    assert expected_vcard_data == vcard
+
+
 def test_vcard_data_invalid_bday():
     with pytest.raises(ValueError):
         helpers.make_vcard_data('Mustermann;Max', 'Max Mustermann', birthday='19760919')
     with pytest.raises(ValueError):
         helpers.make_vcard_data('Mustermann;Max', 'Max Mustermann', birthday='1976-09-19TZ')
+
+
+def test_vcard_data_invalid_geo():
+    with pytest.raises(ValueError):
+        helpers.make_vcard_data('Mustermann;Max', 'Max Mustermann', lat=1.234)
+    with pytest.raises(ValueError):
+        helpers.make_vcard_data('Mustermann;Max', 'Max Mustermann', lng=1.234)
+
+
+def test_vcard_data_valid_geo():
+    expected_vcard_data = 'BEGIN:VCARD\r\nVERSION:3.0\r\nN:Mustermann;Max\r\nFN:Max Mustermann\r\nGEO:46.235197;8.015445\r\nEND:VCARD\r\n'
+    vcard = helpers.make_vcard_data('Mustermann;Max', 'Max Mustermann', lat=46.235197, lng=8.015445)
+    assert expected_vcard_data == vcard
 
 
 def test_vcard_data_invalid_rev():